SVP condition

+++ scripts/fedora.sh
@@ -95,10 +95,20 @@ fi
         echo $'\e[1;31m'$'\e[0m'
         echo $'\e[1;31m'$'\e[0m'
     fi
-    #Installing SVP
+#Installing SVP
+pkgs='/home/carino/SVP\ 4/SVPManager'
+if [which $pkgs &>/dev/null]
+then
     wget https://www.svp-team.com/files/svp4-latest.php?linux
     tar -xf svp4-latest.php?linux
     sudo -u carino ./svp4-linux-64.run && rm svp4*
+else
+    echo $'\e[1;32m'SVP is already installed$'\e[0m'
+    echo $'\e[1;32m'SVP ya está instalado$'\e[0m'
+    echo $'\e[1;32m'SVP уже установлен$'\e[0m'
+    echo $'\e[1;32m'SVP は既にインストールされています。$'\e[0m'
+fi
+    
     #Setting up a hostname
     if [[ $(hostname) == 'fedora' ]];
     then
